Price being calculated correctly by the PricingService. Still need to generate the read model from this new generated event.

+++ source/Conference/Registration/OrderLine.cs
@@ -17,7 +17,7 @@ namespace Registration
 
     public class OrderLine
     {
-        public int LineTotal { get; set; }
+        public decimal LineTotal { get; set; }
     }
 
     public class SeatOrderLine : OrderLine
